requirements.txt
# 核心计算与深度学习
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
einops==0.8.0
numpy==1.26.4
scipy==1.14.0

# 图像处理 & 细胞分割
opencv-python==4.10.0.84
scikit-image==0.24.0
cellpose==3.1.0
tifffile==2024.7.24
imageio==2.34.2
roifile==2024.9.15

# 增强、可视化与合成
imgaug==0.4.0
matplotlib==3.9.2
seaborn==0.13.2
shapely==2.0.6

# YOLO/目标检测支持（如果用到）
ultralytics==8.3.56
pycocotools==2.0.8

# 风格迁移（CycleGAN / CUT 支持）
dominate==2.9.1
pillow==10.0.0
dill==0.3.8

# 工具
pandas==2.2.2
joblib==1.4.2
natsort==8.4.0
tabulate==0.9.0
tqdm==4.65.2
